+++ backend/src/main/java/kernel360/techpick/feature/structure/model/NameProvider.java
@@ -1,0 +1,9 @@
+package kernel360.techpick.feature.structure.model;
+
+// 테스트 코드를 위해 인터페이스를 분리했습니다.
+public interface NameProvider {
+
+	String findPickNameById(Long id);
+
+	String findFolderNameById(Long id);
+}
